chore: update rusqlite from 0.37 to 0.40
Largest change comes from removal of u64 and usize serialization
in rusqlite 0.38.0.
It is possible to enable it back with `fallible_uint` feature
as suggested in https://github.com/rusqlite/rusqlite/pull/1732,
but there is a good reason for this change
as it may result in unexpected failures when
trying to store u64 integers that don't fit in i64,
so I don't enable it back.
With this change conversions between i64 and u64 are explicit.

Using less `usize` is generally good as its size
differs on 32-bit platforms that are rarely tested
but still supported and we have changes that reduce its usage
such as 0defa117a0a302985c61fe211d1a6a948a06d4f6,
014d2ace76790a8484609a4a27fce53ba2fdaba3 etc.

+++ src/sql/migrations.rs
@@ -658,7 +658,7 @@ pub(crate) async fn msgs_to_key_contacts(context: &Context) -> Result<()> {
         return Ok(());
     }
     let trans_fn = |t: &mut rusqlite::Transaction| {
-        let mut first_key_contacts_msg_id: u64 = t
+        let mut first_key_contacts_msg_id: u32 = t
             .query_one(
                 "SELECT CAST(value AS INTEGER) FROM config WHERE keyname='first_key_contacts_msg_id'",
                 (),
@@ -682,10 +682,10 @@ pub(crate) async fn msgs_to_key_contacts(context: &Context) -> Result<()> {
             )
             .context("Prepare stmt")?;
         let msgs_to_migrate = 1000;
-        let mut msgs_migrated: u64 = 0;
+        let mut msgs_migrated: u32 = 0;
         while first_key_contacts_msg_id > 0 && msgs_migrated < msgs_to_migrate {
             let start_msg_id = first_key_contacts_msg_id.saturating_sub(msgs_to_migrate);
-            let cnt: u64 = stmt
+            let cnt: u32 = stmt
                 .execute((start_msg_id, first_key_contacts_msg_id))
                 .context("UPDATE msgs")?
                 .try_into()?;
